refactor(interface-cli): split four command handlers out of main.rs
Extract four self-contained command handlers from the 2375-line
`crates/interface-cli/src/main.rs` into sibling modules, following the
existing `cmd_*.rs` pattern (cmd_account, cmd_backup, cmd_doctor,
cmd_login, cmd_migrate):

- `cmd_review.rs` — `/review` REPL for skill-refinement proposals (~130 lines)
- `cmd_reset.rs` — `assistant reset` wipe + reseed defaults (~60 lines)
- `cmd_persona.rs` — `assistant persona` list/create/use/skill-mode/timeout (~125 lines)
- `cmd_skill.rs` — `assistant skill` list/show/create/delete (~100 lines)

`PersonaCommand`, `SkillCommand`, and `home_agent_root` flip from
private to `pub` in main.rs so the sibling modules can use them.

main.rs shrinks from 2375 → 1965 lines. All 59 lib tests pass; lint
is clean.

First slice of the interface-cli/main.rs split — more sections to
follow (cmd_webui, parse helpers, bootstrap, the REPL itself).
